feat(http): reuse content negotiating handler
