+++ debug/psutil.go
@@ -1,13 +1,10 @@
 package debug
 
 import (
-	"net/http"
+	"context"
 
 	"github.com/alexfalkowski/go-service/encoding"
-	"github.com/alexfalkowski/go-service/errors"
-	nh "github.com/alexfalkowski/go-service/net/http"
 	"github.com/alexfalkowski/go-service/net/http/content"
-	"github.com/alexfalkowski/go-service/net/http/status"
 	"github.com/alexfalkowski/go-service/runtime"
 	"github.com/shirou/gopsutil/v4/cpu"
 	"github.com/shirou/gopsutil/v4/mem"
@@ -16,22 +13,7 @@ import (
 // RegisterPprof for debug.
 func RegisterPsutil(srv *Server, enc *encoding.Map) {
 	mux := srv.ServeMux()
-
-	psutil := func(res http.ResponseWriter, req *http.Request) {
-		ctx := req.Context()
-
-		defer func() {
-			if r := recover(); r != nil {
-				err := errors.Prefix("health", runtime.ConvertRecover(r))
-				nh.WriteError(ctx, res, err, status.Code(err))
-			}
-		}()
-
-		ct := content.NewFromRequest(req)
-		res.Header().Add(content.TypeKey, ct.Media)
-
-		e := ct.Encoder(enc)
-
+	h := content.NewHandler("debug", enc, func(ctx context.Context) any {
 		data := make(map[string]any)
 
 		i, err := cpu.InfoWithContext(ctx)
@@ -60,9 +42,8 @@ func RegisterPsutil(srv *Server, enc *encoding.Map) {
 			"virtual": vm,
 		}
 
-		err = e.Encode(res, data)
-		runtime.Must(err)
-	}
+		return data
+	})
 
-	mux.HandleFunc("/debug/psutil", psutil)
+	mux.HandleFunc("/debug/psutil", h)
 }
